[트립트립] refactor: HOC를 활용한 권한분기
